# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Biblioteca API",
    "description": "API de gerenciamento de biblioteca com OAuth2",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Servidor de desenvolvimento"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "password": {
            "tokenUrl": "/oauth/token",
            "scopes": {}
          }
        }
      }
    },
    "schemas": {
      "Book": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string" },
          "author": { "type": "string" },
          "isbn": { "type": "string" },
          "published_year": { "type": "integer" },
          "quantity": { "type": "integer" },
          "available": { "type": "boolean" }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "username": { "type": "string" },
          "email": { "type": "string" },
          "role": { "type": "string", "enum": ["admin", "user"] }
        }
      },
      "Loan": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "user_id": { "type": "integer" },
          "book_id": { "type": "integer" },
          "loan_date": { "type": "string", "format": "date-time" },
          "due_date": { "type": "string", "format": "date-time" },
          "return_date": { "type": "string", "format": "date-time" },
          "status": { "type": "string", "enum": ["active", "returned", "overdue"] }
        }
      }
    }
  },
  "security": [
    { "bearerAuth": [] }
  ],
  "paths": {
    "/books": {
      "get": {
        "summary": "Lista todos os livros",
        "responses": {
          "200": {
            "description": "Lista de livros",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Book" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Adiciona um novo livro",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Book" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Livro criado"
          },
          "401": {
            "description": "Não autorizado"
          }
        }
      }
    },
    "/books/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "integer" }
        }
      ],
      "get": {
        "summary": "Busca um livro pelo ID",
        "responses": {
          "200": {
            "description": "Detalhes do livro",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Book" }
              }
            }
          },
          "404": {
            "description": "Livro não encontrado"
          }
        }
      },
      "put": {
        "summary": "Atualiza um livro",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Book" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Livro atualizado"
          },
          "401": {
            "description": "Não autorizado"
          },
          "404": {
            "description": "Livro não encontrado"
          }
        }
      },
      "delete": {
        "summary": "Remove um livro",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": {
            "description": "Livro removido"
          },
          "401": {
            "description": "Não autorizado"
          },
          "404": {
            "description": "Livro não encontrado"
          }
        }
      }
    },
    "/loans": {
      "get": {
        "summary": "Lista empréstimos (apenas os do próprio usuário para não-admin)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista de empréstimos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Loan" }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado"
          }
        }
      },
      "post": {
        "summary": "Cria um novo empréstimo",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Loan" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Empréstimo criado"
          },
          "401": {
            "description": "Não autorizado"
          }
        }
      }
    },
    "/oauth/token": {
      "post": {
        "summary": "Obter token de acesso",
        "security": [],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "grant_type": {
                    "type": "string",
                    "enum": ["password", "refresh_token"]
                  },
                  "username": { "type": "string" },
                  "password": { "type": "string" },
                  "refresh_token": { "type": "string" }
                },
                "required": ["grant_type"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token de acesso gerado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": { "type": "string" },
                    "refresh_token": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas"
          }
        }
      }
    }
  }
}