# src/models/bookModel.js
const { DataTypes } = require('sequelize');
const sequelize = require('../config/db.config');

const Book = sequelize.define('Book', {
  id: { type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true },
  title: { type: DataTypes.STRING, allowNull: false },
  author: { type: DataTypes.STRING, allowNull: false },
  isbn: { type: DataTypes.STRING, unique: true },
  published_year: { type: DataTypes.INTEGER },
  quantity: { type: DataTypes.INTEGER, defaultValue: 1 },
  available: { type: DataTypes.BOOLEAN, defaultValue: true }
}, {
  tableName: 'books',
  timestamps: true
});

module.exports = Book;